Added activities and states to different modules for better organization. #2
