+++ client/lib/activities.js
@@ -1,0 +1,15 @@
+const activities = [
+  { name: 'Astronomy', code: 'Astronomy' },
+  { name: 'Biking', code: 'Biking' },
+  { name: 'Hiking', code: 'Hiking' },
+  { name: 'Camping', code: 'Camping' },
+  { name: 'Birdwatching', code: 'Birdwatching' },
+  { name: 'Museum Exhibits', code: 'Museum Exhibits' },
+  { name: 'Fishing', code: 'Fishing' },
+  { name: 'Scenic Driving', code: 'Scenic Driving' },
+  { name: 'Kayaking', code: 'Kayaking' },
+  { name: 'Boating', code: 'Boating' },
+  { name: 'Guided Tours', code: 'Guided Tours' }
+];
+
+module.exports = activities;
